Move common cassandra stuff out of cassandraPersistence
