+++ common/persistence/cassandra/cassandraHelpers.go
@@ -34,8 +34,6 @@ import (
 	"github.com/uber/cadence/tools/common/schema"
 )
 
-const cassandraPersistenceName = "cassandra"
-
 // CreateCassandraKeyspace creates the keyspace using this session for given replica count
 func CreateCassandraKeyspace(s *gocql.Session, keyspace string, replicas int, overwrite bool) (err error) {
 	// if overwrite flag is set, drop the keyspace and create a new one
