+++ common/persistence/dataInterfaces.go
@@ -2422,16 +2422,6 @@ func (config *ClusterReplicationConfig) GetCopy() *ClusterReplicationConfig {
 	return &res
 }
 
-// DBTimestampToUnixNano converts Milliseconds timestamp to UnixNano
-func DBTimestampToUnixNano(milliseconds int64) int64 {
-	return milliseconds * 1000 * 1000 // Milliseconds are 10⁻³, nanoseconds are 10⁻⁹, (-3) - (-9) = 6, so multiply by 10⁶
-}
-
-// UnixNanoToDBTimestamp converts UnixNano to Milliseconds timestamp
-func UnixNanoToDBTimestamp(timestamp int64) int64 {
-	return timestamp / (1000 * 1000) // Milliseconds are 10⁻³, nanoseconds are 10⁻⁹, (-9) - (-3) = -6, so divide by 10⁶
-}
-
 var internalThriftEncoder = codec.NewThriftRWEncoder()
 
 // NewHistoryBranchToken return a new branch token
