+++ common/persistence/persistenceInterface.go
@@ -26,7 +26,6 @@ import (
 
 	workflow "github.com/uber/cadence/.gen/go/shared"
 	"github.com/uber/cadence/common/checksum"
-	"github.com/uber/cadence/common/persistence/managers/shard"
 	"github.com/uber/cadence/common/types"
 )
 
@@ -38,9 +37,6 @@ type (
 	// Manager/TaskManager/MetadataManager are the same.
 	//////////////////////////////////////////////////////////////////////
 
-	// ShardStore is a lower level of Manager
-	// TODO: move this definition to store package and to change references to it then simply remove it in favor of the interface
-	ShardStore = shard.Manager
 	// TaskStore is a lower level of TaskManager
 	TaskStore = TaskManager
 	// MetadataStore is a lower level of MetadataManager
